+++ channels/drive/client/drive_main.c
@@ -1004,7 +1004,7 @@ out_error:
  *
  * @return 0 on success, otherwise a Win32 error code
  */
-UINT drive_DeviceServiceEntry(PDEVICE_SERVICE_ENTRY_POINTS pEntryPoints)
+uintptr_t drive_DeviceServiceEntry(PDEVICE_SERVICE_ENTRY_POINTS pEntryPoints)
 {
 	RDPDR_DRIVE* drive;
 	UINT error;
